backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.1
